requirements.txt
# Hira: Hierarchical Range-Searching Attention

# Core dependencies
torch>=2.0.0
transformers>=4.30.0
numpy>=1.20.0

# Optional dependencies for examples
python-dotenv>=0.19.0
huggingface-hub>=0.10.0

# Optional: FAISS for faster clustering
# faiss-cpu>=1.7.0  # Uncomment for CPU version
# faiss-gpu>=1.7.0  # Uncomment for GPU version

# Development dependencies (install with: pip install -e ".[dev]")
# pytest>=7.0.0
# black>=22.0.0
# flake8>=4.0.0
# mypy>=0.950
